requirements.txt
openai-whisper
sounddevice
soundfile
numpy
fastapi
uvicorn
python-multipart
psycopg[binary]
python-dotenv
